Don't attach custom events to window
