+++ src/libs/sync_manager.js
@@ -1,3 +1,5 @@
+import { listen } from 'src/libs/customEvents';
+
 /**
  Call this function to schedule other functions
  after mapbox Map is loaded.
