+++ src/main.js
@@ -2,7 +2,6 @@ import '@babel/polyfill';
 import 'url-search-params-polyfill';
 import I18n from './libs/i18n';
 import './libs/sync_manager';
-import './libs/listen';
 import App from './panel/app_panel';
 import Store from './adapters/store';
