+++ views/index.ejs
@@ -42,7 +42,7 @@
           data-flag-text="<%= config.app.versionFlag %>"
         <% } %> >
         <input id="search" type="search" class="search_form__input" spellcheck="false" required placeholder="<%= _('Search on Qwant Maps') %>" autocomplete="off">
-        <input type="submit" value="" class="search_form__action" onclick="fire('submit_autocomplete')">
+        <input type="submit" value="" class="search_form__action" onclick="submitSearch()">
         <button id="clear_button" class="search_form__clear icon-x" type="button" onmousedown="clearSearch()"></button>
         <span class="search_form__logo"></span>
       </form><!-- no whitespace
